bakcup

+++ Matrix_Design/Matrix_Design.h
@@ -9,6 +9,7 @@
 #include <iostream>
 #include <ostream>
 #include <type_traits>
+#include <vcruntime.h>
 template <bool B, typename T>
 using Enable_if = typename std::enable_if<B, T>::type;
 
@@ -86,7 +87,7 @@ std::array<size_t, N> computing_stride(const Array &extents) {
   }
 
   for(size_t i = 0; i < N - 1; ++i){
-    int product = 1;
+    size_t product = 1;
     for(size_t j = i + 1; j < N; ++j){
       product *= extents[j];
     }
